# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pulse-brain-lab/lessons:
    get:
      operationId: listPulseBrainLabLessons
      tags: [pulseBrainLab]
      summary: List PulseBrainLab lessons
      description: >
        Lightweight catalog of the curated PulseBrainLab lessons (global
        reference, identical for every school). Optionally filter by grade band.
      parameters:
        - name: gradeBand
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PulseBrainLabGradeBand"
      responses:
        "200":
          description: Lesson summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PulseBrainLabLessonSummary"
  /pulse-brain-lab/lessons/{lessonKey}:
    get:
      operationId: getPulseBrainLabLesson
      tags: [pulseBrainLab]
      summary: Get a single PulseBrainLab lesson
      description: Full lesson object including bilingual parent card and student worksheet.
      parameters:
        - name: lessonKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PulseBrainLabLesson"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PulseBrainLabGradeBand:
      type: string
      enum: ["K-2", "3-5", "6-8", "9-12"]
    PulseBrainLabBrainModelTag:
      type: string
      enum: ["Spotlight", "Velcro", "Echo", "Rewire"]
    PulseBrainLabWorksheetResponseType:
      type: string
      enum: ["write", "draw", "checklist"]
    LocalizedText:
      type: object
      properties:
        en:
          type: string
        es:
          type: string
      required: [en, es]
    PulseBrainLabPrompt:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
      required: [id, text]
    PulseBrainLabLessonFlow:
      type: object
      properties:
        connect:
          type: string
        teach:
          type: string
        practice:
          type: string
        close:
          type: string
      required: [connect, teach, practice, close]
    PulseBrainLabParentReinforcement:
      type: object
      properties:
        summary:
          $ref: "#/components/schemas/LocalizedText"
        brainIdea:
          $ref: "#/components/schemas/PulseBrainLabBrainModelTag"
        askYourChild:
          type: array
          items:
            $ref: "#/components/schemas/LocalizedText"
        whyThisWorks:
          $ref: "#/components/schemas/LocalizedText"
        tryTogether:
          $ref: "#/components/schemas/LocalizedText"
      required: [summary, brainIdea, askYourChild, whyThisWorks, tryTogether]
    PulseBrainLabWorksheetPrompt:
      type: object
      properties:
        id:
          type: string
        text:
          $ref: "#/components/schemas/LocalizedText"
        responseType:
          $ref: "#/components/schemas/PulseBrainLabWorksheetResponseType"
      required: [id, text, responseType]
    PulseBrainLabStudentWorksheet:
      type: object
      properties:
        intro:
          $ref: "#/components/schemas/LocalizedText"
        prompts:
          type: array
          items:
            $ref: "#/components/schemas/PulseBrainLabWorksheetPrompt"
      required: [intro, prompts]
    PulseBrainLabLessonSummary:
      type: object
      properties:
        lessonKey:
          type: string
        gradeBand:
          $ref: "#/components/schemas/PulseBrainLabGradeBand"
        week:
          type: integer
        session:
          type: integer
        title:
          type: string
        skillArea:
          type: string
        brainModelTag:
          $ref: "#/components/schemas/PulseBrainLabBrainModelTag"
      required:
        - lessonKey
        - gradeBand
        - week
        - session
        - title
        - skillArea
        - brainModelTag
    PulseBrainLabLesson:
      type: object
      properties:
        id:
          type: string
        gradeBand:
          $ref: "#/components/schemas/PulseBrainLabGradeBand"
        week:
          type: integer
        session:
          type: integer
        title:
          type: string
        skillArea:
          type: string
        brainConcept:
          type: string
        brainModelTag:
          $ref: "#/components/schemas/PulseBrainLabBrainModelTag"
        objective:
          type: string
        materials:
          type: string
        durationMinutes:
          type: integer
        flow:
          $ref: "#/components/schemas/PulseBrainLabLessonFlow"
        contentQuestions:
          type: array
          items:
            $ref: "#/components/schemas/PulseBrainLabPrompt"
        followupQuestions:
          type: array
          items:
            $ref: "#/components/schemas/PulseBrainLabPrompt"
        skillTags:
          type: array
          items:
            type: string
        parentReinforcement:
          $ref: "#/components/schemas/PulseBrainLabParentReinforcement"
        studentWorksheet:
          $ref: "#/components/schemas/PulseBrainLabStudentWorksheet"
      required:
        - id
        - gradeBand
        - week
        - session
        - title
        - skillArea
        - brainConcept
        - brainModelTag
        - objective
        - materials
        - durationMinutes
        - flow
        - contentQuestions
        - followupQuestions
        - skillTags
        - parentReinforcement
        - studentWorksheet
...