# lib/api-zod/src/generated/types/pulseBrainLabGradeBand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PulseBrainLabGradeBand =
  (typeof PulseBrainLabGradeBand)[keyof typeof PulseBrainLabGradeBand];

export const PulseBrainLabGradeBand = {
  "K-2": "K-2",
  "3-5": "3-5",
  "6-8": "6-8",
  "9-12": "9-12",
} as const;