# lib/api-zod/src/generated/types/pulseBrainLabLessonSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PulseBrainLabBrainModelTag } from "./pulseBrainLabBrainModelTag";
import type { PulseBrainLabGradeBand } from "./pulseBrainLabGradeBand";

export interface PulseBrainLabLessonSummary {
  lessonKey: string;
  gradeBand: PulseBrainLabGradeBand;
  week: number;
  session: number;
  title: string;
  skillArea: string;
  brainModelTag: PulseBrainLabBrainModelTag;
}