# lib/api-zod/src/generated/types/pulseBrainLabWorksheetPrompt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LocalizedText } from "./localizedText";
import type { PulseBrainLabWorksheetResponseType } from "./pulseBrainLabWorksheetResponseType";

export interface PulseBrainLabWorksheetPrompt {
  id: string;
  text: LocalizedText;
  responseType: PulseBrainLabWorksheetResponseType;
}