# lib/api-zod/src/generated/types/pulseBrainLabWorksheetResponseType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PulseBrainLabWorksheetResponseType =
  (typeof PulseBrainLabWorksheetResponseType)[keyof typeof PulseBrainLabWorksheetResponseType];

export const PulseBrainLabWorksheetResponseType = {
  write: "write",
  draw: "draw",
  checklist: "checklist",
} as const;